requirements.txt
google-genai>=1.0.0
twilio>=9.0.0
feedparser>=6.0.10
requests>=2.31.0
